(MODULES-4359) Fix native types so they can be used directly.
This commit makes it possible to use the native types directly without
the defined type wrappers.

The concat_file and concat_fragment have an inexplicable design, rather
than use the :target parameter to associate fragments they each
re-declare the tag parameter as a standard parameter (which afaics does
nothing, i.e. the types behave exactly the same without).

Piggybacking on the tag parameter is also clumsy because its values
are limited to the pattern /\A[0-9A-Za-z_][0-9A-Za-z_:.-]*\Z/R, making
the normalization in 'concat' and 'concat::fragment' kind of required
and using the actual types directly impossible. (At least not without
doing the same everywhere)

There is no explanation on why it was designed this way, leading me to
believe that a) There is something I don't know, b) It's based on a
misconception or c) it just worked and nobody cared.

The defined types are overhead and their raison d'être is to provide
backward compatibility with the old non-native concat implementation
thus they should not act as a required interface to the native types.

The commit makes the following changes while keeping backward
compatibility:

- It drops the explicit tag parameter and associates fragments by their
'target' parameter instead.

- It implements the parameter validations from the defined types directly
in concat_file and concat_fragment.